some seeding

+++ Domain/Inquiry.cs
@@ -2,8 +2,8 @@ namespace Domain;
 public class Inquiry
 {
     public Guid Id { get; set; }
-    public float Width { get; set; }
-    public float Height { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
     public int Weight { get; set; }
     public DateOnly Date { get; set; }
     public Address SourceAddress { get; set; }
